added policy for TeamsController, refs #27289

+++ app/controllers/api/teams_controller.rb
@@ -6,30 +6,23 @@
 #  https://github.com/puzzle/cryptopus.
 
 class Api::TeamsController < ApiController
-  before_action :admin_only, except: :index
+  rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
 
   def index
-    teams = current_user.teams
+    teams = policy_scope Team
     render_json teams
   end
 
   def last_teammember_teams
     teams = user.last_teammember_teams
+    authorize teams
     render_json teams
   end
 
   def destroy
+    authorize team
     team.destroy
     render_json ''
-  end
-
-  protected
-
-  def admin_only
-    unless current_user.admin?
-      add_error t('flashes.admin.admin.no_access')
-      render_json && return
-    end
   end
 
   private
@@ -40,5 +33,10 @@ class Api::TeamsController < ApiController
 
   def team
     @team ||= Team.find(params['id'])
+  end
+
+  def user_not_authorized(_exception)
+    add_error t('flashes.admin.admin.no_access')
+    render_json && return
   end
 end
